myspark/requirements.txt
spark==2.3.0
